hw5-code/pyproject.toml
[tool.poetry]
name = "cs5830_homeworks"
version = "0.1.0"
description = ""
authors = ["CS5830 Staff"]

[tool.poetry.dependencies]
python = "^3.10"
pytest = "^6.2.5"
pycryptodome = "^3.12.0"
black = "*"
requests = "^2.27.1"
numpy = "^1.22.1"
requests-cache = "^0.9.1"
sympy = "^1.10.1"
bitstring = "^3.1.9"
ecdsa = "^0.17.0"
BitVector = "^3.5.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
